requirements.txt
# FastAPI Core
fastapi
uvicorn
python-multipart

# Image Processing
pillow
rembg
opencv-python-headless
pilgram
numpy

# Machine Learning & AI
vertexai
google-auth
scikit-learn
mediapipe==0.10.14
onnxruntime